Moving winOrLose function into separate file
